requirements.txt
django
gunicorn
django-mysql
django-counter-field-py3
reportlab
django-crum
django-versatileimagefield
python-decouple
#django-postgres-extra
weasyprint
#flask
#redis
